feat: Portfolio型の単一ソース化
types/portfolio.ts を作成し Portfolio 型の定義を一元化。
portfolioRepository.ts のインライン型定義を削除し types/portfolio.ts から re-export するよう修正。
